+++ NEWS.md
@@ -1,4 +1,10 @@
+# BGmisc beta 1.5.2
+* More flexible ID generation for simulatePedigree
+* Created ped2gen function to extract generation information from pedigree data.frames
+* Added tests for ped2gen
+
 # BGmisc 1.5.1
+## CRAN submission
 * partially refactored summarizePedigree to be more modular
 * added compression control to ped2com
 * Minor copy editing